// src/api/openapi/auto-generated/models/DocsAIChatConversationModel.ts
export interface DocsAIChatConversationModel {
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatConversationModel
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatConversationModel
     */
    customerId: string;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatConversationModel
     */
    userId: string;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatConversationModel
     */
    sessionId: string;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatConversationModel
     */
    questionId: string;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatConversationModel
     */
    question: string;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatConversationModel
     */
    answer: string;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatConversationModel
     */
    prompt?: string;
    /**
     * 
     * @type {Date}
     * @memberof DocsAIChatConversationModel
     */
    askedAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof DocsAIChatConversationModel
     */
    respondedAt: Date;
    /**
     * 
     * @type {DocsAIChatFeedbackModel}
     * @memberof DocsAIChatConversationModel
     */
    feedback?: DocsAIChatFeedbackModel;
    /**
     * 
     * @type {DocsAIChatOptionsSetModel}
     * @memberof DocsAIChatConversationModel
     */
    optionsSet?: DocsAIChatOptionsSetModel;
    /**
     * 
     * @type {Date}
     * @memberof DocsAIChatConversationModel
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof DocsAIChatConversationModel
     */
    updatedAt: Date;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatConversationModel
     */
    createdBy: string;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatConversationModel
     */
    updatedBy: string;
}