// src/api/openapi/auto-generated/models/DocsAIChatFeedbackModel.ts
export interface DocsAIChatFeedbackModel {
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatFeedbackModel
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatFeedbackModel
     */
    customerId: string;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatFeedbackModel
     */
    userId: string;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatFeedbackModel
     */
    conversationId: string;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatFeedbackModel
     */
    feedback?: string;
    /**
     * 
     * @type {number}
     * @memberof DocsAIChatFeedbackModel
     */
    thumbAction?: DocsAIChatFeedbackModelThumbActionEnum;
    /**
     * 
     * @type {Date}
     * @memberof DocsAIChatFeedbackModel
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof DocsAIChatFeedbackModel
     */
    updatedAt: Date;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatFeedbackModel
     */
    createdBy: string;
    /**
     * 
     * @type {string}
     * @memberof DocsAIChatFeedbackModel
     */
    updatedBy: string;
}


/**
 * @export
 */
export const DocsAIChatFeedbackModelThumbActionEnum = {
    NUMBER_MINUS_1: -1,
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_unknown_default_open_api: 11184809
} as const;
export type DocsAIChatFeedbackModelThumbActionEnum = typeof DocsAIChatFeedbackModelThumbActionEnum[keyof typeof DocsAIChatFeedbackModelThumbActionEnum];